wrote tests that are passing

+++ OptionTool.Domain/Entities/OptionInformation.cs
@@ -5,7 +5,7 @@
         public string OptionId { get; set; }
         public string OptionSequence { get; set; }
         public string Label { get; set; }
-        public long? TranslationId { get; set; }
+        public decimal? TranslationId { get; set; }
         public OptionValueType OptionValueType { get; set; }
         public ValueLookupObject ValueLookupObject { get; set; }
         public ValueEnumGroup ValueEnumGroup { get; set; }
